backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
sqlalchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.1
nltk==3.8.1
spacy==3.7.2
numpy==1.26.3
pandas==2.2.0
python-docx==1.1.0
PyPDF2==3.0.1
scikit-learn==1.4.0
python-magic==0.4.27
pytest==8.0.0
httpx==0.26.0